+++ packages/core/plop/component/index.js
@@ -53,13 +53,8 @@ module.exports = plop => {
       },
       {
         type: "add",
-        path: "src/components/{{properCase componentName}}/__tests__/{{camelCase componentName}}-snapshot-tests.jest.js",
-        templateFile: "plop/general/component-snapshot-tests-jest.txt"
-      },
-      {
-        type: "add",
-        path: "src/components/{{properCase componentName}}/__tests__/{{camelCase componentName}}-tests.jest.js",
-        templateFile: "plop/general/component-tests-jest.txt"
+        path: "src/components/{{properCase componentName}}/__tests__/{{properCase componentName}}.test.js",
+        templateFile: "plop/general/component-tests.txt"
       },
       {
         type: "append",