+++ packages/core/src/components/EditableHeading/__stories__/EditableHeading.stories.js
@@ -21,12 +21,18 @@ const editableHeadingTemplate = createComponentTemplate(EditableHeading);
 
 export const Overview = {
   render: editableHeadingTemplate.bind({}),
-  name: "Overview",
   args: {
     value: "This heading is an editable heading",
     type: EditableHeading.types.H1
   },
-  play: overviewPlaySuite
+  play: overviewPlaySuite,
+  parameters: {
+    docs: {
+      liveEdit: {
+        isEnabled: false
+      }
+    }
+  }
 };
 
 export const Types = {
@@ -94,5 +100,11 @@ export const Types = {
       </div>
     </div>
   ),
-  name: "Types"
+  parameters: {
+    docs: {
+      liveEdit: {
+        scope: { EditableHeading }
+      }
+    }
+  }
 };