+++ packages/core/src/components/Search/Search.tsx
@@ -19,6 +19,7 @@ const Search = forwardRef(
       clearIconName = CloseSmallIcon,
       clearIconLabel = "Clear",
       renderAction: RenderAction,
+      hideRenderActionOnInput,
       value,
       placeholder,
       size = "medium",
@@ -90,7 +91,7 @@ const Search = forwardRef(
           />
         )}
         {inputValue && !disabled && ClearIcon}
-        {RenderAction}
+        {!(hideRenderActionOnInput && inputValue) && RenderAction}
       </>
     );