+++ packages/storybook-blocks/.eslintrc.js
@@ -39,7 +39,7 @@ const commonExtends = ['plugin:react/recommended', 'plugin:prettier/recommended'
 module.exports = {
   overrides: [
     {
-      files: ['*.jest.js', 'jest.setup.js', 'jest.init.js'],
+      files: ['*.test.js', 'jest.init.js'],
       env: {
         jest: true,
         'jest/globals': true,
@@ -76,7 +76,7 @@ module.exports = {
   },
   settings: {
     jest: {
-      version: 27,
+      version: 29,
     },
     react: {
       version: 'detect',